# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "An API to get cocktails and ingredients",
        "title": "CocktailGrid API",
        "contact": {},
        "version": "1"
    },
    "basePath": "/",
    "paths": {
...
        "/cocktails/{cocktailId}/ingredients/{ingredientId}": {
            "post": {
                "description": "Adds an existing ingredient to the cocktail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cocktails"
                ],
                "summary": "Adds an ingredient",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cocktail ID",
                        "name": "cocktailId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Ingredient ID",
                        "name": "ingredientId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Cocktail ingredient object",
                        "name": "cocktail",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CocktailIngredientDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.CocktailIngredientResultDto"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dtos.CocktailIngredientDto": {
            "type": "object",
            "properties": {
                "quantity": {
                    "type": "number"
                }
            }
        },
        "dtos.CocktailIngredientResultDto": {
            "type": "object",
            "properties": {
                "cocktailId": {
                    "type": "integer"
                },
                "ingredientId": {
                    "type": "integer"
                },
                "ingredientName": {
                    "type": "string"
                },
                "quantity": {
                    "type": "number"
                }
            }
        },
...
    }
}